[REFACTOR] 프로젝트/패키지명 변경
