+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'today-escape'
+rootProject.name = 'o-escape'
