+++ src/main/java/com/nextroom/oescape/controller/TestController.java
@@ -1,4 +1,4 @@
-package com.todayescape.controller;
+package com.nextroom.oescape.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
