+++ src/main/java/com/nextroom/oescape/domain/Hint.java
@@ -1,4 +1,4 @@
-package com.todayescape.domain;
+package com.nextroom.oescape.domain;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
