+++ src/main/java/com/nextroom/oescape/exceptions/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package com.todayescape.exceptions;
+package com.nextroom.oescape.exceptions;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
