+++ src/main/java/com/nextroom/oescape/exceptions/StatusCode.java
@@ -1,4 +1,4 @@
-package com.todayescape.exceptions;
+package com.nextroom.oescape.exceptions;
 
 import org.springframework.http.HttpStatus;
