+++ src/test/java/com/nextroom/oescape/OEscapeApplicationTests.java
@@ -1,10 +1,10 @@
-package com.todayescape;
+package com.nextroom.oescape;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class TodayEscapeApplicationTests {
+class OEscapeApplicationTests {
 
     @Test
     void contextLoads() {
